+++ config/passport.js
@@ -4,7 +4,6 @@ const { getAccount } = require("../db/queries");
 const bcrypt = require("bcryptjs");
 
 const verifyCallback = async (username, password, done) => {
-  console.log("verifyCallback running...");
   try {
     const account = await getAccount({ username, password });
     const match = account && (await bcrypt.compare(password, account.password));
@@ -23,12 +22,10 @@ const strategy = new LocalStrategy(verifyCallback);
 passport.use(strategy);
 
 passport.serializeUser((account, done) => {
-  console.log("serializeUser running...");
   done(null, account.id);
 });
 
 passport.deserializeUser(async (id, done) => {
-  console.log("deserializeUser running...");
   try {
     const account = await getAccount({ id });
     return done(null, account);