+++ controllers/accountController.js
@@ -8,7 +8,6 @@ const { createAccount, getAccount } = require("../db/queries");
 
 const accountController = {
   getLogin: asyncHandler(async (req, res) => {
-    console.log("getLogin running...");
     res.render("login");
   }),
   getLogout: asyncHandler(async (req, res) => {
@@ -16,8 +15,6 @@ const accountController = {
     res.redirect("/");
   }),
   getSignup: asyncHandler(async (req, res) => {
-    console.log("getSignup running...");
-    console.log("req.user:", req.user);
     res.render("signup");
   }),
   postLogin: [
@@ -26,8 +23,6 @@ const accountController = {
       const errors = validationResult(req);
       const inputs = matchedData(req, { onlyValidData: false });
       if (!errors.isEmpty()) {
-        console.log(errors.mapped());
-        console.log("inputs:", inputs);
         return res.render("login", {
           errors: errors.mapped(),
           inputs,
@@ -37,13 +32,7 @@ const accountController = {
       next();
     }),
     (req, res, next) => {
-      console.log("postLogin running...");
-      console.log("req.body:", req.body);
       passport.authenticate("local", (err, account, info) => {
-        console.log("authenticating...");
-        console.log("err:", err);
-        console.log("account:", account);
-        console.log("info:", info);
         if (err) return next(err);
         if (!account) {
           res.locals.message = info.message;
@@ -52,17 +41,14 @@ const accountController = {
           return res.status(422).render("login");
         }
 
-        console.log("req.login called...");
         return req.login(account, next);
       })(req, res, next);
     },
     (req, res) => {
-      console.log("postLogin running after authentication....");
       res.redirect("/my-account");
     },
   ],
   postLogout: (req, res, next) => {
-    console.log("postLogout running...");
     req.logout((err) => {
       if (err) return next(err);
       res.redirect("/");
@@ -71,13 +57,10 @@ const accountController = {
   postSignup: [
     validateSignup,
     asyncHandler(async (req, res, next) => {
-      console.log("postSignup running after validateAccount...");
       const errors = validationResult(req);
       const inputs = matchedData(req, { onlyValidData: false });
 
       if (!errors.isEmpty()) {
-        console.log("errors.mapped():", errors.mapped());
-        console.log("inputs:", inputs);
         return res.render("signup", {
           errors: errors.mapped(),
           inputs,
@@ -100,11 +83,9 @@ const accountController = {
         });
 
         const account = await getAccount({ username });
-        console.log("account:", account);
 
         // Automatically login after creating account
         req.login(account, (err) => {
-          console.log("login running after creating account...");
           res.redirect("/");
         });
       } catch (err) {