+++ controllers/indexController.js
@@ -5,26 +5,19 @@ const { insertPost, getAllPosts, deletePost } = require("../db/queries");
 
 const indexController = {
   setAllPosts: asyncHandler(async (req, res, next) => {
-    console.log("setAllPosts running...");
     const posts = await getAllPosts();
-    console.log("posts:", posts);
     res.locals.posts = posts;
     next();
   }),
   getIndex: asyncHandler(async (req, res) => {
-    console.log("getIndex running...");
-    console.log("req.url", req.url);
     res.render("index");
   }),
   postIndex: [
     validatePost,
     asyncHandler(async (req, res, next) => {
-      console.log("postIndex running...");
       const errors = validationResult(req);
       const inputs = matchedData(req, { onlyValidData: false });
       if (!errors.isEmpty()) {
-        console.log(errors.mapped());
-        console.log("inputs:", inputs);
         return res.render("index", {
           errors: errors.mapped(),
           inputs,
@@ -41,9 +34,7 @@ const indexController = {
     }),
   ],
   postDeletePost: asyncHandler(async (req, res) => {
-    console.log("postDeletePost running...");
     const { postID, accountID } = req.params;
-    console.log("req.params:", req.params);
     await deletePost({ accountID, postID });
     res.redirect("/");
   }),