+++ routes/myAccountRouter.js
@@ -9,7 +9,6 @@ const {
 const myAccountRouter = new Router();
 
 const isAuthenticated = (req, res, next) => {
-  console.log("isAuthenticated running...");
   // If an account is logged in, redirect to the root
   if (!req.isAuthenticated()) return next({ status: 401 });
   next();